Remove the typecast (int64 to int32) in fused_experts_cpu

+++ python/sglang/srt/layers/moe/topk.py
@@ -697,7 +697,7 @@ def fused_topk_torch_native(
 
     if renormalize:
         topk_weights = topk_weights / topk_weights.sum(dim=-1, keepdim=True)
-    return topk_weights, topk_ids
+    return topk_weights, topk_ids.to(torch.int32)
 
 
 def fused_topk_softmax_torch_raw_logits(
